Separate tool definitions into different json files. Remove hardcoded segments related to gto. Move tool categories to the json definition file.
The tool definitions are now located in /public/tools and have a tool index in /public that lists them. This index is fetched by the application to know which tools exist.

The tools are now loaded when the `AllOperationsPanel` is created. This required changing how they were "obtained" in the code so i created a `toolsUtils.js` with the proper getters and changed the places in the code that were referencing the `descriptions.json` file.