simplify the loading and running of the wasm modules
previously we were creating individual wrappers to create the wasm modules and return the function to run them. i have changed this to be a single function to run an individual tool from the js generated by emscripten.

also changed it so the user input is passed directly to the stdin so that we dont need the main_wrapper.c

with this come changes to the bash script to compile the tools since we dont need the wrappers anymore